implement display and edit as separate screens
