+++ messages/en.json
@@ -393,22 +393,25 @@
   "screens.Security.title": {
     "message": "Security"
   },
-  "screens.Setting.ProjectSettings.DeviceName.deviceNameDisplayHeader": {
+  "screens.Setting.ProjectSettings.DeviceName.DisplayScreen.deviceNameLabel": {
     "message": "Your Device Name"
   },
-  "screens.Setting.ProjectSettings.DeviceName.deviceNameEditHeader": {
+  "screens.Setting.ProjectSettings.DeviceName.DisplayScreen.title": {
+    "message": "Device Name"
+  },
+  "screens.Setting.ProjectSettings.DeviceName.EditScreen.deviceNameLabel": {
     "message": "Edit Device Name"
   },
-  "screens.Setting.ProjectSettings.DeviceName.discardChangesAlertCancelText": {
+  "screens.Setting.ProjectSettings.DeviceName.EditScreen.discardChangesAlertCancelText": {
     "message": "Continue Editing"
   },
-  "screens.Setting.ProjectSettings.DeviceName.discardChangesAlertConfirmText": {
+  "screens.Setting.ProjectSettings.DeviceName.EditScreen.discardChangesAlertConfirmText": {
     "message": "Discard Changes"
   },
-  "screens.Setting.ProjectSettings.DeviceName.discardChangesAlertTitle": {
+  "screens.Setting.ProjectSettings.DeviceName.EditScreen.discardChangesAlertTitle": {
     "message": "Discard changes?"
   },
-  "screens.Setting.ProjectSettings.DeviceName.title": {
+  "screens.Setting.ProjectSettings.DeviceName.EditScreen.title": {
     "message": "Device Name"
   },
   "screens.Setting.ProjectSettings.YourTeam.InviteAccepted.addAnotherDevice": {
